First part of dijkstra implementation

+++ src/dataStructures/graph/Edge.java
@@ -1,4 +1,4 @@
 package dataStructures.graph;
 
-public record Edge<K extends Comparable<K>, E>(Vertex<K, E> start, Vertex<K, E> destination, double weight) {
+public record Edge<K extends Comparable<K>, E>(Vertex<K, E> start, Vertex<K, E> destination, int weight) {
 }
